Move assignments loading inside model
